* Removed several GCC 4.3.x compilation warnings, including some paranoid   ones. Removed several unused variables, and optimized the memory layout   of a few classes.

+++ exult/trunk/files/Table.h
@@ -55,7 +55,7 @@ public:
 		: U7file(spec)
 		{  }
 
-		virtual uint32 number_of_objects(void)
+		virtual size_t number_of_objects(void)
 			{ return object_list.size(); };
 		virtual char *retrieve(uint32 objnum,std::size_t &len);
 		virtual const char *get_archive_type()
